don't add any APK, instead delete unknown APKs that don't match metadata
The current process of searching for APKs is imperfect, so it often grabs
random APKs that should not be included.  This script was then dutifully
creating metadata for those new APKs and adding them to the feed.  Instead,
manually add the metadata and delete any unknown APKs to keep this repo
clean.